Refactor: organiza componentes y carpetas
