+++ src/components/pages/home/SocialMediaGallery.jsx
