+++ src/components/pages/login/Login.jsx
@@ -1,7 +1,7 @@
 // src/components/pages/Login.jsx
 import React from "react";
-import LoginBackground from "../login/LoginBackground"; // Asegúrate de que coincida con la estructura real
-import LoginForm from "../login/LoginForm"; 
+import LoginBackground from "./LoginBackground"; // Asegúrate de que coincida con la estructura real
+import LoginForm from "./LoginForm"; 
 
 function Login() {
   // Podríamos ocultar Navbar, SubNav y Footer para esta página
